requirements.txt
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0
python-dotenv>=1.0.1
undetected-chromedriver>=3.5.5
webdriver-manager>=4.0.1
